// mumuca-bass-worker/src/modules/track/interfaces/metadata.ts
export interface Metadata {
  title: string;
  artist: string;
  album?: string;
  year?: number;
  trackNumber?: number;
  genre?: string;
  composer?: string;
  lyricist?: string;
  producer?: string;
  arranger?: string;
  performer?: string;
  comment: string;
  isrc?: string;
  coverArtwork?: string;
  copyright?: string;
  discNumber?: number;
}